requirements.txt
django-crispy-forms
crispy-bootstrap5
requests-oauthlib
python3-openid
anyascii==0.3.2
arrow==1.3.0
asgiref==3.8.1
beautifulsoup4==4.12.3
binaryornot==0.4.4
certifi==2024.7.4
cffi==1.17.0
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
cryptography==43.0.0
cssselect2==0.7.0
cssutils==2.11.1
defusedxml==0.7.1
Django==5.0.8
django-allauth[mfa,socialaccount]>=65.0.2
qrcode >= 7.0.0
django-classy-tags==4.1.0
django-cogwheels==0.3
django-debug-toolbar==4.4.6
django-entangled==0.5.4
django-extensions==3.2.3
django-filer==3.2.0
django-filter==24.3
django-formtools==2.5.1
django-inlinecss==0.3.0
django-js-asset==2.2.0
django-ldap==0.0.8
django-modelcluster==6.3
django-mptt==0.16.0
django-parler==2.3
django-permissionedforms==0.1
django-polymorphic==3.1.0
django-rest-framework==0.1.0
django-sekizai==4.1.0
django-taggit==5.0.1
django-treebeard==4.7.1
djangorestframework==3.15.2
draftjs-exporter==5.0.0
easy-thumbnails==2.9
elastic-transport==8.15.0
elasticsearch==8.15.0
et-xmlfile==1.1.0
filetype==1.2.0
flake8==7.1.1
future==1.0.0
html5lib==1.1
idna==3.8
iniconfig==2.0.0
jwt==1.3.1
l18n==2021.3
laces==0.1.1
ldap3==2.9.1
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
more-itertools==10.4.0
npm==0.1.1
openpyxl==3.1.5
optional-django==0.1.0
packaging==24.1
pluggy==1.5.0
psycopg2-binary==2.9.9
pyasn1==0.6.0
pycodestyle==2.12.1
pycparser==2.22
pyflakes==3.2.0
Pygments==2.18.0
pynliner==0.8.0
pytest==8.3.2
python-dateutil==2.9.0.post0
python-slugify==8.0.4
python-webpack-boilerplate==1.0.3
pytz==2024.1
PyYAML==6.0.2
reportlab==4.2.2
requests==2.31.0
setuptools==73.0.1
soupsieve==2.6
sqlparse==0.5.1
svglib==1.5.1
telepath==0.3.1
text-unidecode==1.3
types-python-dateutil==2.9.0.20240821
urllib3==2.2.2
Willow==1.8.0
whitenoise>=5.3.0
toml>=0.10.2
Unipath>=1.1
python-decouple>=3.4
gunicorn>=20.1.0
autopep8>=1.5.7
dj-database-url>=0.5.0
django-compressor
django-cors-headers>=4.4.0
pillow>=9.3.0
pyjwt
python3-saml
